Rename root level `api` folder to `apiFunctions`
